Move layout and charts into multiple modules, replace PerformanceReport class
