+++ README.md
@@ -12,7 +12,7 @@ https://stock-analysis-web-app.herokuapp.com/
 
 > Note: The app is hosted on a small (free!) dyno hosted by Heroku. If the app doesn't work, try again later or refresh the page a few times until the dyno has woken up again ;).
 
-![gif](./assets/web-app.gif)
+![gif](./web-app.gif)
 
 
 ### Running locally
