+++ requirements.txt
@@ -6,7 +6,6 @@ Flask==2.2.2
 gunicorn==20.1.0
 matplotlib==3.6.0
 numpy==1.23.3
-pandas==1.5.0
+pandas==1.4.0
 plotly==5.10.0
-seaborn==0.12.0
 xlsxwriter
